traefik-authproxy/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-jose==3.5.0
requests==2.32.4
PyYAML==6.0.2
